Refactor handling of pre-validations in business request handler
